+++ src/main/java/insane96mcp/experienceoverhaul/mixin/client/AnvilScreenMixin.java
@@ -1,5 +1,6 @@
 package insane96mcp.experienceoverhaul.mixin.client;
 
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import insane96mcp.experienceoverhaul.module.anvil.AnvilXpCost;
 import insane96mcp.insanelib.core.feature.Feature;
 import net.minecraft.client.gui.GuiGraphics;
@@ -16,9 +17,7 @@ import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Constant;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyConstant;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @OnlyIn(Dist.CLIENT)
@@ -33,24 +32,17 @@ public abstract class AnvilScreenMixin extends ItemCombinerScreen<AnvilMenu> {
         super(menu, inventory, title, background);
     }
 
-    @ModifyConstant(method = "renderLabels", constant = @Constant(intValue = 40))
+    @ModifyExpressionValue(method = "renderLabels", at = @At(value = "CONSTANT", args = "intValue=40"))
     private int experienceoverhaul$tooExpensiveCap(int cap) {
         if (!Feature.isEnabled(AnvilXpCost.class)) return cap;
         return AnvilXpCost.repairCap;
     }
 
-    @Inject(method = "renderLabels", at = @At("TAIL"))
-    public void experienceoverhaul$renderFreeLabel(GuiGraphics guiGraphics, int mouseX, int mouseY, CallbackInfo ci) {
-        if (this.menu.getCost() != 0 || !this.menu.getSlot(2).hasItem()) return;
+    @Inject(method = "renderLabels", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/screens/inventory/ItemCombinerScreen;renderLabels(Lnet/minecraft/client/gui/GuiGraphics;II)V", shift = At.Shift.AFTER), cancellable = true)
+    public void experienceoverhaul$preventLabelRender(GuiGraphics guiGraphics, int mouseX, int mouseY, CallbackInfo ci) {
+        if (this.menu.getCost() != 0 || !this.menu.getSlot(2).hasItem())
+            return;
 
-        int color = 8453920;
-        Component label = Component.translatable("container.repair.free");
-        if (!this.menu.getSlot(2).mayPickup(this.player))
-            color = 16736352;
-
-        int x = this.imageWidth - 8 - this.font.width(label) - 2;
-        int y = 69;
-        guiGraphics.fill(x - 2, 67, this.imageWidth - 8, 79, 1325400064);
-        guiGraphics.drawString(this.font, label, x, y, color);
+        ci.cancel();
     }
 }