+++ changelog.md
@@ -1,4 +1,4 @@
-## 1.0.1.1
+## Upcoming
 * Fixed mod not required client side
 
 ## 1.0.1.0-alpha