+++ changelog.md
@@ -2,7 +2,7 @@
 * Fixed mod not required client side
 
 ## 1.0.1.0-alpha
-* Added back gamerule to disable experience
+* Added back game rule to disable experience
 
 ## 1.0.0.0-alpha
 * Initial release