+++ changelog.md
@@ -1,4 +1,7 @@
 ## Upcoming
+* Ported Anvil features
+  * Anvil Degradation
+    * Change degradation chance and allow repairing with iron blocks (configurable with item tag)
 * Fixed mod not required client side
 
 ## 1.0.1.0-alpha