+++ changelog.md
@@ -1,4 +1,4 @@
-## Upcoming
+## 1.1.1.0
 * Removed the game rule to disable experience
   * Moved to Rune Enchanting
 * Cost label is no longer shown if cost = 0 instead of showing No Cost